Fix task deletion :bug:

+++ modules/deleteData.js
@@ -1,16 +1,20 @@
 const deleteData = (app, passport, io, id) => {
   const DBquery = require('./DBquery');
   const getUsers = require('./getUsers');
-
+  const notify = require('./notify');
 
   io.on('delete data', data => {
-    if (data.key !== undefined) {
+    if (data.key !== undefined && data.key !== '') {
       query = `DELETE FROM ${data.table} WHERE ${id} = '${data.key}'`;
 
       DBquery(app, io, 'DELETE FROM', data.table, {
           text: query
         })
-        .then(() => getUsers(app, passport));
+        .then(() => {
+          if (data.table === 'users') getUsers(app, passport);
+        });
+    } else {
+      notify(io, 'failure');
     }
   });
 }
